Python Piscine/p01/requirements.txt
numpy
flake8
Pillow
matplotlib